backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pdf2image>=1.17.0
google-genai>=1.0.0
ollama>=0.4.0
reportlab>=4.2.0
redis>=5.0.0
pillow>=10.0.0
sqlalchemy>=2.0.0
